requirements.txt
python-telegram-bot==13.15
openai==1.30.1
python-dotenv==1.0.1
filetype==1.2.0
